requirements.txt
aiohttp==3.8.4
beautifulsoup4==4.10.0
Google_Images_Search==1.4.6
imageai==3.0.3
Pillow==9.5.0
python-dotenv==1.0.0
RealESRGAN==1.0
RealESRGAN==0.3.0
Requests==2.31.0
torch==2.0.1+cu118